[Driver][Python] Detection and reallocation for contiguous tensors (#45510)
This PR introduces the ability to determine if a tensor is contiguous
and a method to create a contiguous copy of a tensor. This also makes a
change to the API to reject execution of any noncontiguous tensors and
errors informing the user to reallocate a tensor in a contiguous
fashion.

Closes MSDK-785

MAX_PYTHON_ORIG_REV_ID: 1b07dbbfff37bc6c5733449e1acfe3be495696df