[mojo-test] Allow --filter with --debug (#45613)
Adds support for `--filter` when `--debug` is passed, to restrict the
set of tests being debugged. This required a change to the entrypoint's
interface; rather than accepting a test ID as an argument it accepts a
file path containing a newline-separated list of test IDs to run. This
is because the maximum length of a command-line invocation is limited;
for large codebases with many tests, we could easily exceed that length
without it being obvious to the user.

Fixes MOTO-659.

MODULAR_ORIG_COMMIT_REV_ID: edc57dab2b2122e60a623b02abb341e974d53ebc