[mojo-vscode] RPC call for cuda-gdb mode, prompt user to install nsight (#45580)
The cuda-gdb debugging mode used by `mojo debug --rpc --cuda-gdb`
requires the nsight-vscode-edition extension to vscode. We don't want to
make that extension a required dependency of ours, but we require it for
this particular feature. This PR adds checking for the extension, and
prompts the user to install it if not found.

Stacked on [Internal link]

MODULAR_ORIG_COMMIT_REV_ID: 83f0d9b057ee3411a0a43d3649aeb7807af530bb