[KGEN] Continue simplifying interpreter APIs (NFC?) (#45740)
This removes the requirement that a stack frame is active to make stack
allocations. This means stack allocations created before a frame is
active are "immortal" with respect to the execution of a function. This
is technically more correct, because references to stack allocations
passed into the entry function live in a caller's context, not inside
the entry function as we have been treating it until now. It's not clear
if this has any functional effects. It probably makes returning a
reference to an argument work.

This allows hiding all the stack frame fiddling inside
`callFunctionBody` and `interpretFunction` (renamed from
`runInterpreter`). This moves towards adding a bytecode backend for the
interpreter.

MODULAR_ORIG_COMMIT_REV_ID: 7f100a48026bee520eb26d35aac53df4baf0e1b5