[vscode] Remove the 'run in terminal' action (#45718)
This action is all wrong. It executes `mojo` in the current terminal,
but in order for `mojo` to work, the `MODULAR_HOME` env var needs to be
set. If the current terminal doesn't have that env var, the `mojo`
invocation fails. Instead, we should only keep the `run in dedicated
terminal` action, which creates a brand new terminal for each file and
properly injects the required env var. This will ensure that the `debug`
and `run` actions behave the same way.

MODULAR_ORIG_COMMIT_REV_ID: 912a825d449a7ae3bc474fa057a91042261ab7ba