[vscode] Make the extension look like nightly by default (#45719)
An annoyance that we currently have is that whenever we want to run the
vscode extension in dev mode, we need to disable the current nightly
extension. The reason is that the vscode extension is by default the
stable one, and you can't have both the stable and the nightly at the
same time.
A simple way to fix this is to make the extension look like the nightly
by default, and only before publishing we can change it to look like the
stable one.

MODULAR_ORIG_COMMIT_REV_ID: 85c77d47ec9c8cbbd1069213c8ffeaa8fbc18718