[ADT] Introduce `ReallocatingTrivialVector` (#45802)
This is a vector type of POD element types. It uses `realloc` to grow
and provides direct, unsafe access to the underlying data. This is
useful for reducing memory pressure when re-using a vector of POD
elements. For example, a `SmallVector<Attribute>` might be better used
as a `ReallocatingTrivialVector<Attribute>`.

MODULAR_ORIG_COMMIT_REV_ID: 99830aa10f38c0ed2cd5fe49ab433a6db3060fa3