[KGEN] Improve task management in elaborator (NFC) (#45851)
This PR jiggles how task management works in the elaborator. This is an
attempt to get better thread utilization by ensuring tasks on the
critical path of the program are executed. This also removes the caching
step of `compilerElaboratorAsm`. It's not clear if we need to cache
at this granularity across compiler processes.

MODULAR_ORIG_COMMIT_REV_ID: 3d9913734c7b8f23e9a316476bac948041982193