[Max Engine] Use Driver Device. (#45849)
We had 2 APIs to set devices in the engine for various reasons. This PR
removes the old one and replaces the uses with Device APIs.

### Things changed
- Removed old API and migrated its uses to new API.
- Removed GPU execution support from modular-api-executor. We don't plan
to support this at C API for now.
- Disabled executing models in GPU. This was the old path and the tests
will be re-enabled soon when we have support for GPU tensors in driver.

MAX_PYTHON_ORIG_REV_ID: 48301317b748009fbb14d4398a810c9629d31253