[mojo-lang] Add failed attempt to improve diagnostic. (#45922)
We currently print the implicit parameters on types with them.  The
problem is that the metatype to the StructType is wrong: it doesn't
have the pogs for the metatype, so we can't know what is an implicit
parameter.  This PR adds some dead code and a testcase for the
issue.

MODULAR_ORIG_COMMIT_REV_ID: bc57dac4fb31a7e61ca10381db95da023de4584a