[vscode] Add an spec for the behavior of the extension (#45928)
I realized that the extension is getting rather complex, and that it's
better to document how it should behave, so that we can implement this
spec and add proper tests. So far this has more than 200 lines, so it's
definitely not simple.

MODULAR_ORIG_COMMIT_REV_ID: 75c3c09f94445d587b21101fe1995b61a550aff1