[MO] Remove AddDevicesPass and DeviceSpecMap (#43901)
This pass and type are no longer used, so remove them. Instead of check
with the runtime to find the available devices, we use the
`AssignDevicesPass` and take input from options to determine which
device we should assign to the top-level graph up.

While this is not a non-functional change, it really should not impact
the user-facing behavior in any cases.

Towards GRA-658

MODULAR_ORIG_COMMIT_REV_ID: 4b4822ccf3f41bb7cabfd2255b90402f73e301cc