[Kernels] migrate optimizations warp->smem layer_norm kernel (#45987)
Change the shared memory version of the layer_norm kernel to also use a
float type for intermediates. The two versions should be in sync now.

Fix the dispatch logic to not use the warp version if the column count
is not a multiple of simd_width. The warp kernel can run with
simd_width=1, but that also means that the column count range is
reduced. The column count for typical models in practice is a multiple
of simd_width, so instead of complicating the dispatch logic, just
always fallback to the shared memory version.
NN_ORIG_REV_ID: 321d470cc5cb3817f01400e51aab3c55fac6180a