[vscode] Reimplement SDK management for modular cli and dev SDKs (#46001)
This implements most of the parts of behavior spec that correspond to the modualr cli and dev SDKs. The greatest benefit is that almost nothing is cached, therefore the extension will always be in sync with the SDKs.

MODULAR_ORIG_COMMIT_REV_ID: c3f2b2256aed0ecbf717a356575ea31d09838a3a