[mojo-lang] Hill climb field sensitive lifetime infra. (#46011)
This adds `__lifetime_of` and ASTType printing for `LifetimeFieldAttr`.
The former is a bit gross because we cannot just emit the base
expression
and check to see if it is an MValue: this doesn't work in dynamic
contexts
like argument lists.

This includes an unrelated NFC refactor of `emitStoredFieldRef` as well.

MODULAR_ORIG_COMMIT_REV_ID: 6fa9e33b77f766bd2f5750801c8d70a4ee99368a