[serialization] [mlir] MLIR de/serialization in mojo (#45294)
Uses the `str` representation of a graph to serialize a graph to MLIR
and then reload the MLIR to create a graph object which should be the
same as the original. This will be unified with the mef serialization to
provide a portable way to serialize models.

---------

Signed-off-by: Rashid Kaleem <rashid@modular.com>
Co-authored-by: Fehr Mathieu <mathieu.fehr@gmail.com>
Co-authored-by: Arjun Surendran <arjun@modular.com>
MAX_INTEGRATION_TESTS_REV_ID: 3bc0ede1f543977c7d2eac98b49b8f693423cb27