[vscode] Run integration tests in CI (#45861)
This adds a new GHA workflow for running the Mojo VS Code extension's
integration tests against PRs.

This would, in a vacuum, be better implemented as a Bazel target.
However, it's possible we will move the extension out of this repository
in the future, so having a workflow we can carry over when that happens
will also be helpful. It also means we don't need to make changes to
Bazel or the extension, which requires a `modular.cfg` file (environment
variable overrides are not supported).

MODULAR_ORIG_COMMIT_REV_ID: 9ca669e66b56175e8884371e83214690865cc655