[bazel] Build tcmalloc on macOS (#46145)
There are some issues integrating tcmalloc's usptream build in bazel,
specifically it depends on autoreconf from the system, which we are
silently using on Linux, but not macOS.

This prebuilds tcmalloc for macOS with `MACOSX_DEPLOYMENT_TARGET=11.0
build-tcmalloc` and vendors the result. Since we don't change our
patches or update this that often this should be fine for now.

The ideal here is that we write a BUILD file that works for macOS and
Linux

MODULAR_ORIG_COMMIT_REV_ID: 860c89eb211eda01a49e4be24293dd3f68641f8f