[vscode] Add a script that allows you to invoke the mojo installed by the extension (#46166)
If magic is enabled in the extension, the extension installs
automatically max and mojo. This means that there's in the system a very
stable mojo that can be invoked without rebuilding anything. A side
effect is that the debugger can be used for this.
With this change, if you do `vscode-mojo debug --vscode <target>`, the
extension will use the debugger from the magic installation.

MODULAR_ORIG_COMMIT_REV_ID: a1c02b4f4d9cd9b9cf4e89d0f13d83f762cf6086