[mojo-lang] Change the syntax of `lit.ref.struct.ger`, ~NFC (#46208)
This changes the syntax of the `lit.ref.struct.ger` to start with
the type and lifetime of the base, and follow with the resultant
element type.  This is more logical, and also makes way for field
sensitive lifetimes, where the result lifetime is more complex
(and derived from) the source lifetime.

MODULAR_ORIG_COMMIT_REV_ID: 5a38ecdc79a3f57c4d5989453f31bc4d3f9e0d4c