[mojo-lang] Teach `#lit.lifetime.union` about subtyping. (#46222)
This teaches `#lit.lifetime.union` to fold elements in a union when
they are subsets of other elements.  Specifically, a reference to
lifetime `x->y` is redundant with a lifetime of `x` so we can retain
just the second when placed in a union.

This is the key for making lifetime subtyping work, which is an
important part of field sensitive lifetimes.

MODULAR_ORIG_COMMIT_REV_ID: 901a9c64ac5c5893be3db5c180646890bbd9054b