[bazel] Improve lit test output (#45877)
Previously lit tests worked by appending custom substitutions to tools,
since they are not found in the default lit paths. Now we just fixup the
substitutions the user added and make sure they are in the PATH of the
test run.

This also means that the paths to the binaries in the lit output are no
longer full bazel paths, but instead are relative paths that are
copy-pastable for use with the //:install target.

Unfortunately the paths to the lit test files were also absolute paths
so this overrides the test formatter to hook the output and replace the
bazel paths with the relative paths in the repo.

With these changes many lit commands are copy pastable 1:1. This also
removes the need for some custom_substitutions that were just there to
avoid things looking in the PATH.

MODULAR_ORIG_COMMIT_REV_ID: f9d228f5cd4e09106ab8bdbcced118f735c1190e