[KGEN][CompilationServer] Roundtrip LLVM module (#46144)
Adds functionality to pass MLIR module as a string to `emitArchive`,
parse it into an op and send it back as a string. This is an enabler to
build full implementation of `emitArchive` on the server. Once that's
functional, the return string will contain emitted archive rather than a
copy of the compiled module. Full implementation will not use this
functionality at all and will pass input and output of `emitArchive`
through memory.

---------

Co-authored-by: Tatiana Shpeisman <tatiana@Tatianas-MBP.attlocal.net>
Co-authored-by: Jeff Niu <jeff@modular.com>
MODULAR_ORIG_COMMIT_REV_ID: 3f7d7550879a36e6c8a7fe53b254a233a2de5dee