[vscode] Add a mutex in the extension activation (#46307)
We have different code paths that trigger reactivation of the extension,
so it's possible that there's a race condition and two of them run at
the same time. In order to prevent this, we can add a mutex.

MODULAR_ORIG_COMMIT_REV_ID: 78541990dc447ccc2a4c1d4720dab19edce4ad0d