[KGEN] Add lowering for `pop.noalias_pointer_cast` (#46377)
This lowers it to a custom KGEN LLVM intrinsic that has a `noalias`
result. This tells LLVM's basic AA that the pointer doesn't locally
alias with anything else.

The body is `return %arg0` with the appropriate attributes. It's marked
as alwaysinline, which is how the function gets "lowered".

MODULAR_ORIG_COMMIT_REV_ID: fd4d46eb3cb7a08617863696abaf129182a8664a