[Build] Introduce a new kernel_e2e_gpu_profiling build mode (#46402)
This will be used in subsequent PRs to allow us to propagate more kernel
information to the nsight traces to get a better understanding of the
kernels being dispatched. Note that unlike the profiling mode, this does
not turn on llcl profiling, since it adds overhead. It also keeps the
build in production mode which is important so that we do not measure
added overhead.

MODULAR_ORIG_COMMIT_REV_ID: 9b92a467672690cdb69e35be4f04331c95181b2f