[Kernels][GPU] Rework how function caching is done (#46420)
A bunch of the patterns that we were using was because we had to go
through a c++ layer in the past. We no longer have this limitation, so
can operate on a faster pathway.

GPU_ORIG_REV_ID: 6cf4c9ee792f904b5d3aba2c1c77288f5580ec86