[mojo-lang] Make unpack logic simpler and more robust (#46459)
This changes the implementation of `*_` to expand until it fills as many
positional only or posOrKw parameters the caller requires. This also
means it can expand to empty and overwrite values passed as keywords.
This matches Python's unpacking semantics more and also is a more robust
implementation.

This came as I was adding `**_` and found a bunch of issues with the
current implementation. This will compose better with unpacking.

MODULAR_ORIG_COMMIT_REV_ID: 2036bf388cb8d4ae91dff008f5a80961f7e7f2b2