[Kernels] Fix GPU `layer_norm` to use biased sample variance (#46530)
When I initially ported the GPU `layer_norm` tests to the CPU, the tests
did not pass and that was because the CPU and GPU use a slightly
different formulation for the variance calculation inside the kernel.
The GPU was using the unbiased sample variance (`mean / (n - 1)`) but
the CPU and all other implementations I could find uses the biased
sample variance (`mean / n`). This fixes the GPU kernel and test to use
the same formulation.

NN_ORIG_REV_ID: 19ac5dcd536e4cf71efc69db6e8744a65d78c591