[Kernels] export `rms_norm` kernel as a custom op (#46606)
This completes the kernel side of exposing the `rms_norm` custom op for
use by llama3 models.

The custom op needs to export the scalar `epsilon` parameter as a
`NDBuffer`. Ideally, this would still be a Scalar, but the graph
compiler handles these scalar parameters by extracting the value from a
managed tensor and this managed tensor is placed on the GPU. There isn't
sufficient machinery exported in the SDK to access the device placement
attribute of a custom op.

The `layer_norm` registration has been fully moved to
nn/normalization.mojo to mirror the handling of `rms_norm`.

The export is tested by the SDK pipelines and that change will be
submitted after this lands.

MOGG_ORIG_REV_ID: 04af942cfc1b9658800dd0b078584730e88d514a