[Kernels][Pipelines] Added KV Cache Kernels for Tiny Llama (#46633)
For some of our testing, we are using a TinyLlama variant, which uses 1
kv head, with a head dim of 10. This adds support for this configuration
in advance of moving the Python Llama3 pipeline to use the mo.opaque KV
Cache.

Fixes MSDk-905

MOGG_ORIG_REV_ID: eebd533dce573618e419d168cdc371203f5a295b