[mojo-lang] Verify that signatures with capture lifetimes are closures (#46711)
It would only make sense for a function to have capture lifetimes if it
is a closure, i.e. marked as 'capturing'. Add a verifier to check for
this until there is a reason to loosen it.

MODULAR_ORIG_COMMIT_REV_ID: 5cd0da58008a8815757f4045a7d9ac19abe6825c