[KGEN] Move canonicalization registration to register-custom-ops (#46769)
Canonicalization patterns were JIT'ed in `canonicalizer`, which was
quite complex as we had to avoid with data races. JIT'ing the patterns
in `register-custom-ops` makes it easier.

The drawback is that the JIT'ed canonicalization patterns are not in the
IR anymore, making it harder to debug.

Co-authored-by: Mathieu Fehr <mathieu.fehr@gmail.com>
MODULAR_ORIG_COMMIT_REV_ID: fbd156a9e5b1af705dca44fce9b4de853d867211