[mojo-lang] Make field sensitive lifetimes refuse to collapse.
This changes lifetime tracking to consider "x.y" a distinct element from
"x" and
not collapse it. The rationale for this is that (with the previous
behavior) we have
two different kinds of subfield references, "x.y" could be an use of
"the y field of x"
or it could be a use of "the y field of something embedded in x" which
are different.

It is possible that we might want to enable something like this as sugar
in the future,
but that would require introducing syntax (e.g. "x.y" vs "x.~y") or
whatever, and I'd
prefer to keep the model simple, rather than adding frills.

This makes use of the previous patch's extension to make typeof and
lifetimeof
much more general, which makes it possible to do this without massive
suffering :-)

MODULAR_ORIG_COMMIT_REV_ID: 63d33858d7f797a1537ba1bf58fc4abe2ed26347