The `#lit.lifetime` attribute was added early in the development of Mojo lifetimes, representing a singleton lifetime used for lowering. As things evolved, it got confused semantics: is it the "everything" lifetime, or the "nothing" lifetime, or the "global" lifetime?  Mojo is all over the place on this.
This leads to a number of complications and some weird behavior, and we
need
to clarify this. This patch takes a step forward on this, renaming it to
`#lit.any.lifetime`, specifying that it can alias any other lifetime,
and fixing our
subtyping rules around lifetime union collapsing etc.

As part of this, we make a tiny user visible change to diagnostics, and
introduce
new `ImmutableAnyLifetime` and `MutableAnyLifetime` which more precisely
capture these effects. This retains `*utableStaticLifetime` for now, but
its
behavior should be refined with an "actually globally static" lifetime
some day.

This patch mostly rearranges things without changing behavior in a
significant
way. Notably, this still allows implicit conversion from references of
"any" lifetime
to a specific lifetime, even though that isn't correct. This will be
refined more
in the subsequent steps.

MODULAR_ORIG_COMMIT_REV_ID: 069dce44898815cf476bafb826396f5f4193699f