[mojo-lang] Improve exclusivity errors. (#46800)
This improves the error message to include the callee name when
available and generate a note when generated from a synthesized
method like a trait stub.  This also improves the location info
for those stubs to correspond to the original method instead of
the location of the struct enclosing them.

MODULAR_ORIG_COMMIT_REV_ID: fb99b0e531e214c2e8d30e158f65afcd4463fc87