[GenericML] Replace mogg mlirbc with shared library (#46098)
These targets produce files that take a huge amount of time for clang to
build, when all we want is to store opaque data for access at runtime.
Because of this we split them out, but this meant that anything that
used them needed to have them set in their data. This caused stale file
issues at runtime when they were missing from data but included from a
previous run of the install target.

Now we embed these as objects in a shared library and load from that
library at runtime.

MODULAR_ORIG_COMMIT_REV_ID: d9c9e8e29a7851197403a57f645dd690b48893fb