[KGEN] Massively reduce surface area of custom ops feature (#46896)
The implementation of "custom ops", which are meant for library-based
rewrites, current take up a lot of surface area. The feature is also not
production ready. This PR massively reduces the surface area of the
feature throughout the codebase, especially the frontend and in the pass
pipeline, and pushes as much of the non-trivial code as possible into a
single MLIR pass.

This PR removes pretty much all the frontend changes, for example.

This should hopefully reduce the maintenance burden for now so that one
can mostly forget about this.

MODULAR_ORIG_COMMIT_REV_ID: ce0889109f4b7257899968f2af2efc502f155a6f