[stdlib] Restore implicit `Copyable` for `Tuple`, `ListLiteral`, etc
While we'd like to remove the implicit copyability requirement in the long-term
for many core standard library types, we're in a bit of a "stuck" state due to
some compiler bugs with traits and conditional conformance to keep pursuing this
work.  So, unwind this work a bit to get out of this halfway state we're
currently stuck in by re-introducing the implicit copyability requirement (via
`CollectionElement` trait) to several types like `Tuple` and `ListLiteral`.

In the future, we can revisit things like `InlineArray` that currrently *only*
work on `CollectionElementNew` to allow them to work with `CollectionElement`
again until we pick this work back up.

EXTENSIBILITY_ORIG_REV_ID: 43863eb6e3faee2355615f8c90a1ed2409cd87c3