[Kernels] Switch kernels to use OptionalReg which is unchecked Optional (#47134)
Optional is checked which is not useful for the kernels which alraedy
check values, so there is a double check. Optional also has behavior
which is undefined on GPUs, so stop using Optional in the kernel code
and switch many places to use OptionalReg. A followup would introduce an
UncheckedOptional which we'd use in more places.

GPU_ORIG_REV_ID: c1bc7b07a6bc812a15d140e972882965f3258005