[mojo-stdlib] Change the API of TaskGroup lifetime set (NFC) (#47142)
This moves the lifetime set to the `wait` function instead of putting it
on the type. This is important because the exclusivity checker will
confuse the lifetime set on the taskgroup with the one one the
coroutine.

RUNTIME_ORIG_REV_ID: 191b0ba21a17dd27735cc72f88480a7bcea90a11