[bazel] Use binary id for kgen cache (#45961)
The linker embeds an id based on the content of the binary. This id can
be read at runtime instead of the previous files based method in order
to provide a unique ID suitable for the mojo cache. This is much faster
than hashing the binary ourselves, and since it's content based should
be perfect for this use.

MODULAR_ORIG_COMMIT_REV_ID: 40b04da4da97539ec91640f223923b230779c261