[lit] Remove cuda and has_cuda_device features (#47279)
In bazel we have to include this same info in the BUILD file so that
bazel knows what machines to send the test to with remote execution.
Because of this we were just always enabling these features, so they're
unnecessary now.

ASYNCRT_ORIG_REV_ID: f6520941d7bbd97f13cf75e85e6db2f0154e9e56