[All] Step 1 of many to support FP8 in our stack (#47404)
This is needed for H100 support, so start prefetching the work to add
support for FP8 now so that this is ready ahead. This just add the basic
datastrctures and dispatches in the compiler. There is going to be a
bunch more changes (in followup PRs) to enable this to work in the
graph.

This is also skipping NUZ variants of FP8

MODULAR_ORIG_COMMIT_REV_ID: 0c81d92b26ef40f02e01f6983a7f2e39916d8b99