[KGEN] Generalize compile_assembly emission kind (#47452)
Fixes
[Internal link]

Allow the emission kind field in a compile_assembly operator to be
non-immediate. Fallback to a generic param value if the emission kind is
not an immediate constant. To disambiguate the immediates with param
values of the same names, immediates are now prepended with '='.

MODULAR_ORIG_COMMIT_REV_ID: a67c5652f6fd94d9a1c8df017a51ce94e3e825b3