[Kernels][GPU] Refactor test_matmul (#47198)
Stacked PRs:
 * __->__#47198

--- --- ---

### [Kernels][GPU] Refactor test_matmul

This PR tries to remove duplication in the `test_matmul.mojo`. The
purpose of PR is to trigger discussion on the design of the test driver
so we will have a robust test for matmul.

The original idea is to pass a lambda with only updates the value of
output matrix given an index. This implementation has a second buffer to
store random values which will be applied to a value of the output
matrix.

@asaadaldien proposes using index to generate an unique value in the
epilogue. It uses a lambda which exactly matches the signature of the
real epilogue passed to the matmul kernel. The PR is adapted to this
interface. This implementation is simpler than previous one.

Both cases have pros and cons.

The first approach is commented out in the PR.

INTERNAL_UTILS_ORIG_REV_ID: 6c05b1db79a97bbc3c4a2c2d6e979ac2c90e17d2