[SDK][Driver] Implement dlpack for `MemMapTensor` (#47528)
The trick here is that since we rely on an `mmap` in the implementation,
we need to ensure that it doesn't get freed too early. We achieve this
by introducing a keyword-only `_mmap` argument on our `__dlpack__`
implementation, pass the `mmap`, and manually bump its refcount (and
then decrement it in the deleter callback).

MAX_TESTS_TESTS_ORIG_REV_ID: f17a303a19070b39fd822b097bfada533c69f014