[KGEN] Update param type asm sugar (NFC) (#47522)
KGEN IR used to have syntax sugar for a "ParamType", which is a type,
that, when it's a parameter, we want to sugar. This means it had two
cases:
1. The type is a ParamRefType wrapping a parameter. We want to just
print the parameter.
2. The type is any other mlir Type. We print the whole thing.

Previously, we handled both cases in one line, by wrapping the type in a
TypeConstantAttr so that it's always a parameter, we can just use the
parameter value printing sugar instead. Case 1 would cancel out the
ParamRef and reveal the original parameter. Case 2 would just be a
simple TypeConstantAttr of a mlir Type, which just prints the mlir Type.

However, this didn't work so well if, in case 1, the parameter didn't
have a TypeType. Since technically we would not be cancelling anything
out, but casting the parameter to a different metatype. This was
revealed by [Internal link] when casting
metatype is properly preserved. This caused the sugar to break.

This PR restores the sugar by explicitly handling the cases:
1. The type is a ParamRefType wrapping a parameter. We just print the
parameter by getting the parameter.
2. The type is any other mlir Type (incl. TypeValueType). We print the
whole thing.

MODULAR_ORIG_COMMIT_REV_ID: 83026ac5886151bee485602eb399f81be79bd386