[mojo-lang][KGEN] Disable a flaky test. (#47644)
This test is flaking in both CI and for me locally.  I'm not sure
what is going on, but we can't have that, so I disabled it and
filed MOCO-1221 to track what is going on.

MODULAR_ORIG_COMMIT_REV_ID: 799212aeb44dfb4d1d4f56e95401e22c66e1ae10