[MGP] Remove cudaRuntime from mgp.debug.tensor.print.cuda (#47515)
This removes the last use of the CUDARuntime.

Previously mgp.debug.tensor.print.cuda would allocate a host buffer and
do a d2h copy inside the primitive, and then print the host buffer,
relying on the CUDARuntime to do so. In order to remove the CUDARuntime
dependency, mgp.debug.tensor.print.cuda now only prints the tensor
metadata, and the mgp author must manually do the alloc + d2h in order
to print the contents of the host buffer.

mo.debug.tensor.print gets lowered to a host alloc + d2h + sync + host
print in mgp so that we can still easily inspect the contents of a
tensor in the graph.

When we have access to the device context in C++ we could restore the
mgp primitive behavior to the way it was, but this unblocks some other
cleanup in the meantime.

Fixes GRA-1009.

---------

Co-authored-by: Ian Tramble <ian@modular.com>
MOJO_MAX_ORIG_REV_ID: a6c09168e3f050e02e91b1c066f0d420039a017a