[mojo-lang] Enhance CheckLifetimes to treat 'AnyLifetime' as uses.
We recently defined the `#lit.any.lifetime` lifetime, which stands in
for a use of
an unknown lifetime that could be "anything".  This commonly occurs when
lifetimes get erased by unsafe constructs.

This changes CheckLifetimes to treat operations that use this lifetime
as
implicitly use of ALL of the tracked values in the current scope that
are live.
This is implemented in two phases: 1) uninit value tracking notices this
situation, and looks at what values are live, and adds an attribute to
the
op in question. 2) dtor insertion then uses this liveness information to
know
what is "known live" and treats them as a use.

The net effect of this, is that Mojo will be a lot more friendly in the
face of
unsafe types, at least those that have `#lit.any.lifetime` lifetime in
them.
I will follow up with a stdlib change to actually take advange of this.

MODULAR_ORIG_COMMIT_REV_ID: 2b25e6b9abf9de8f45d4c1066d270a4a7b9c2cd0