[mojo-lang] Remove SignatureType Opaque Pointers from Blobs (#47243)
This patch replaces the opaque pointer entry with an coordinate entry
into a symbol table I added to the MemorySpace. In internalization, this
coordinate is mapped to an index to a table of symbols maintained by the
interpreter. In externalization these indices are remapped out to a
coordinate to a symbol table built from internal symbols.

MODULAR_ORIG_COMMIT_REV_ID: 5d18aad5a91f7bb04c6855ff3219a615a3e2cb13