[Kernels][GPU] Allow one to specify the runtime path for asm output (#47755)
This enables one to provide a parameter function to the dump methods to
enable one to capture runtime values and put them in the path. This is
useful to dump the asm for all ops of interest into distinct files that
contain other metadata like the shapes.

Right now the passed in function generates the Path, but another design
point (which we can switch to based on use) is providing a callback
function that would pass in the ptx and the callback's responsibility is
to write the file.

Closes KERN-965

GPU_ORIG_REV_ID: 8d15aa0e0528f62f8db0fbaa60aa2d8ed057af8c