[DebugInfo] push ptx name sanitization to debuginfo (#47790)
We have special name mangling for symbols for the PTX backend, but we
previously did not push this mangling through the associated debuginfo,
which breaks cuda-gdb debugging. Well, it blocks debugging after other
features that are still missing are hacked in. But this is a critical
component for end-to-end register debugging.

Closes
[Internal link]

MODULAR_ORIG_COMMIT_REV_ID: 16e201309414c932360099aa719127a0fd54776d