[mojo-lang] Don't mark `ref` args with parametric mutability noalias (#47896)
This refines #47892 to be more conservative and align with exclusivity
checking.  If you have a function that takes an argument as a ref with
parametric mutability, it can be instantiated with an immutable ref and
exclusivity checking will allow it to alias other immutable reference
arguments.  As such, we can't mark it noalias.

This changes the compiler to only mark `ref` arguments with known
mutability as `noalias`.  In practice, it won't matter anyway, because
parametric mut references cannot be mutated.

MODULAR_ORIG_COMMIT_REV_ID: c4a4ab1cf1f2a2979913be07c7c0a799ad492ca1