[kernels] Fix null pointer bug in apple_accelerate.mojo (#48040)
There is a bug where reassigning an `UnsafePointer`, and then using that
to also reassign an `NDBuffer` results in an invalid pointer that is
null, or has an odd address. This resolves that by putting the pointer in
the outer scope, and only allocating when transposing at runtime.

Issue for the bug here:
[Internal link]

Try reverting when this lands: [Internal link]

LINALG_ORIG_REV_ID: c49d0a79c747d3b72d1b08f027a6631c30a44a9d