[mojo-lang] Fix a rather horrific bug in CheckLifetimes. (#47876)
When reinitializing an UnsafePointer, we can end up in a situation
where it looks like we're interacting with AnyLifetime, but also
initializing a value.  In this case, we cannot treat the call as
being a USE of the value getting initialized - such a thing will
get lifetime markers wrong and lead to LLVM miscompiling the code
in pretty subtle and horrifying ways.

MODULAR_ORIG_COMMIT_REV_ID: bb6946da53df8ddd51f9b8173f451561bf86aa78