[mojo-lang] Fix a bug in CheckLifetimes (#48062)
This fixes MOCO-1251, where a lifetime reference to a subfield of
a value didn't extend the use to the entire field.  While here,
unify some logic and change the "full object" comparison to be
based on a ValueRef instead of a Value.  This has the effect of
allowing us to elide a bunch of unneeded reloads of register boxes.

MODULAR_ORIG_COMMIT_REV_ID: 6ba597adfcbacae2e520c135d3cc48b55f3cb6cf