[Kernels] change kv_cache flash_attention to use `scale` parameter (#48132)
The flash attention kernels have been ignoring the supplied scale
parameter and has been hardcoding the `isqrt(head_size)` expression
internally. That was done because the kernels wanted the scale to be
placed on the host, but that was not possible before the fix for #47165
landed.

GPU_TEST_ORIG_REV_ID: 9bcfbab2b9134b86d45d0ced1f6861c33e58029f