[Benchmarking] Turn off padding in SoL model; use better numbers for vLLM (#48324)
The SoL light model shouldn't include padding; that was a bug and this
fixes it. We may re-enable padding when testing other scenarios of our
own.

The old vLLM numbers resulted in a paper-model throughput that was too
low -- even w/o serving overhead, the throughput wasn't as high as
observed. Looking at the data we've collected, we think there are some
scheduling overheads in the benchmark_latency.py measurements that
shouldn't be included in the fake model. This uses different numbers to
try to remove those overheads. The result is a paper speed-of-light
around 17.5 req/s, which is at least above the observed 16 req/s.

We should continue to refine these measurements, but this is definitely
better than before.

SERVE_ORIG_REV_ID: 02cb8c2deeb6c763cef398486049e48eae1ea744