[graph-api]  Don't skip the test_store_slice_buffer test. (#48282)
[graph-api]  Don't skip the test_store_slice_buffer test.

The bug was in using the index 1 rather than 0 in the test case. When creating buffers as test inputs, we
filter out dims of size 0 but not those of size 1. index 0 is not out of range but 1 is.

closes MSDK-1072

MAX_GRAPH_API_ORIG_REV_ID: 876fe11332fa50913c9726e568217797f742bc70