[Support] Add a function to wait for a debugger (#48216)
This function prints the process executable path and the process ID, then
wait for a given number of seconds before resuming executable. It is
designed to allow a debugger to be attached while it is waiting.

This behaves differently from the existing function
`waitForDebuggerToAttach` in `KGEN/lib/Support/Debugging.cpp`, as it
does not raise a `SIGSTOP` signal. Doing so prevents lldb from attaching
properly in some cases.

MODULAR_ORIG_COMMIT_REV_ID: 56ee85b4117caa760ccf3537cbc947960998ef22