[GPU CI] Clean-up A100 GPU tests 2nd try (#48452)
Adding A100 tag for tests that we want to selectively run on A100.
The main motivation here is that A100 is a very expensive CI machine and
many testing can be covered with other GPU instances (A10 for example)
Will use A100 tag in addition to the gpu tag for tests that will run on
A100 in addition to the default gpu.
Also disabled some tests that we do not want to run at all as they do
not add value.

MAX_INTEGRATION_TESTS_REV_ID: 9ac4249376aaddca84569da52b057ad133fd3313