[MOGG][Extensibility] Add support for lambdas_have_fusion KGEN kernel parameter (#48430)
[MOGG][Extensibility] Add support for lambdas_have_fusion KGEN kernel parameter

Add support for the lambdas_have_fusion parameter needed for the
convolution kernels
It's initialized to false, but changed to true if the convolution is
fused with another kernel.
- Automatically had the KGEN parameter for lambdas_have_fusion in
MoToMOGG (set to false).
- Create a new helper to update the KGEN parameters (to factorize with
the existing code in MarkTrivialKernels that update the synchronous
parameter)
- Update FuseInOutLambda options to be able to load correctly the
MojoLibraryAnalysis (needed to update the parameters)
- Update FuseInOutLambda to set lambdas_have_fusion to true after fusion
- Update MOGGAnnonate and OutlineMOGGFuncs to ensure the
lambdas_have_fusion infos is forwarded all the way to the mogg.kernel
IR.
- Add unit units for lambdas_have_fusion with and without fusion

MODULAR_ORIG_COMMIT_REV_ID: efbb70fff27208107efea8d22f181999b5ecbb4f