[GenericML] Improve error message for constraints (#48508)
When a custom op or kernel has a failure in a Mojo `constrained`
statement, format that error message along with the final error that the
graph compiler code will report.

Previously this error message appeared, but could end up buried in a
wall of text dumped to standard output. With this change, the error
message is part of the failure message from the graph compiler, so the
caller can display it immediately to the user.

Fixes MSDK-993

MAX_INTEGRATION_TESTS_REV_ID: 3dba9a32ae1cf475b0215e3523a0614f3a9a48d1