[stdlib] `debug_assert` on GPU show block and thread idx
Makes it so a failed `debug_assert` in the context of a GPU ****** will
stop execution and show an error message with the block and thread idx.
This behaves the same way as `assert` when compiling with nvcc.

The limitation is that the message currently only works with a single
`StringLiteral`.

MODULAR_ORIG_COMMIT_REV_ID: 8d934c0afb63f7e354f147644b70e4626af8217f