[Kernels][Build] Ensure the lit config propagates arch_name constraint (#48686)
This makes sure that the lit config adds the arch_name as a constraint
to lit. This allows us to delete a bit of code that was specific to top
models and correctly constraint other code which is only applicable on
H100.

---------

Co-authored-by: Keith Smiley <keithbsmiley@gmail.com>
GPU_TEST_ORIG_REV_ID: 9c823ee06293014fd80a981b90858bad331a8c51