[mojo-lang] Search Keyword Only Defaults when CTAD Arg Is Unbound (#48743)
The parameters of an enclosing struct are prepended to methods before
binding verification. The default argument information is discarded
during this process. The ParamBindings class stores the immediate
parent's default arguments as a workaround to this lost information so
that unspecified parameters with default values could be inferred. The
problem is that this was only implemented for position parameter
arguments, not keyword only parameter arguments. This patch addresses
this limitation.

MODULAR_ORIG_COMMIT_REV_ID: be8a262686b3a137c7ef33bb4c6bbb4d6c19a7a8