[Kernels][GPU] Add an eviction policy flag to the copy_async functions (#48859)
This adds a copy async flag to the copy_async function. This allows one
to be able to copy the data but mark it with policies like evict first
which is great in cases where the data is mostly read only.

GPU_ORIG_REV_ID: d96255c9151d6d896244101fa2c7cab33d84c270