[Support] Remove nvtx from the TimeProfiler (#49001)
The nvtx profiler in the TimeProfiler did not work (or only worked in
certain cenarios) and instead we have pushed this functionality down to
the mojo runtime library. This removes the special nvtx build along with
the header files that we were depending on.

MODULAR_ORIG_COMMIT_REV_ID: 99b7193764c45a540b60a573c0ff2b8f9e592d34