[Kernels] Start adding gpu back to back matmul. (#48415)
This implements a back to back gemm using `multistage_mma`. The implementation follows the algorithm of flash attention 2 (minus the online softmax), and is thus like a simplified version of that from `mha.mojo`. I started trying to write it independently, but I ended up relying on that file heavily during debugging, for examples of correct use.

Calculates `D = A@B@C`
Assuming
A is M x K
B is K x L
C is L x N
D is M x N
The implementation relies on `N` being known at compile time, and `block_tile_N = warp_tile_N = N`.

`block_tile_N` is currently necessity, as the implementation currently avoids writing `A@B` to memory, and thus recalculates it for every block tile across `N` (meaning we need there to be only 1).
`warp_tile_N` is because a reduction across warps hasn't implemented.

In theory, we could support `N > block_tile_N` via not holding the accumulator of `$(A@B)@C` in registers. Every time we calculate a tile of `A@B`, we could then sweep the columns of `D` and `C` to update them all with the in-register `A@B` block.
This doesn't seem like a win vs non-fused matmuls, as our `$(A@B)@C` would now load from two arrays (C and D) and write to one (D) for every tile of `A@B`, which should increase memory bandwidth requirements by adding the need for writes. More precisely, for every time of `D`, successive matmuls needs to load `L/BK` tiles from `A@B` and `C`, and write `1` tile of `D`. With back to back where do don't keep the accumulator in registers, we load that tile and the corresponding tile from `C` a total of `L/BN` times, and write `L/BN` times. Perhaps this could still be a win, given `BN > BK`.

By keeping the accumulation of `D` in registers, we do have a clear advantage, only needing to load from `C` (although we do need to load the entire `C` per tile of `A@B`).

Tasks:
- [x] Switch to using `multistage_mma`
- [x] Add tests
- [x] Debug
- [ ] Profile and benchmark
- [ ] Improve perf

---------

Co-authored-by: abdul dakkak <adakkak@modular.com>
GPU_TEST_ORIG_REV_ID: fd9c3cc2f3548f7dc1f4653999101464618fc971