[DebugInfo][GPU] disable dbgValueToDeclare completely for GPU (#48806)
This disables convertDbgValueToDeclare for GPUs, which removes
allocation hacks for GPUs. Originally I was going to delete the
conversion completely, but it turns out that it is necessary for certain
cases on CPUs. It should still be scaled back, since it grew a lot to
support GPUs. I'll rework my previous version of this PR into something
that does this conversion in fewer cases to support specific CPU needs
(which can also be used on GPU as desired, but provide less value with
more risk for GPU).

Depends on two more of my upstream PRs being included in an LLVM
integrate:
* https://github.com/llvm/llvm-project/pull/111456
* https://github.com/llvm/llvm-project/pull/111596

Closes
[Internal link]

This is also the last sub-issue in the enclosing epic:
[Internal link]

MODULAR_ORIG_COMMIT_REV_ID: 8fa4f74c7ae5b054c4cb8e06b202d8efb2aa24d1