[KGEN][MOGGPreElab] Support Annotating Needed Parameters For StaticTuple (#48926)
We choose to represent variadic inputs as `StaticTuples` of `ManagedTensorSlice`. This is analogous to what is done in NDBuffer for [Internal link]

For example: a `%variadic_out1, %variadic_out2 = mo.foo(%normal_arg, %variadic_in1, %variadic_in2, %variadic_in3)` will have inputs which map to a `StaticTuple` with a static size of `3`, and an output which maps to `StaticTuple` with static size of `2`. Right now, we assume all tensors in the tuple will have the same rank and dtype.

This PR adds support for annotating these types with parameters that will need to be inferred by the graph compiler from MO/MOGG graphs.

MODULAR_ORIG_COMMIT_REV_ID: 0c701d7c5b31036fa1e8ba33113b86a7927d11fb