[Kernels][GPU] Perform vectorized dot product in the mha kernel (#49109)
Switches the dot operation in the mha kernel to be vectorized. When
measuring the perf, there is no performance improvements are observed,
but there might be an oportunity for tuning. This code is exersized for

```
        test[
            DType.float32,
            DType.float32,
            482,
            32,
            group=4,
        ](482, 482, ctx, is_benchmark())
```

when used in test_mha_causal_mask.mojo

NN_ORIG_REV_ID: 0602431c64ec13cddcc6fd2fd490ce49b96ff6c3