[KVCache][Pipelines] Fuse KVCacheCollection ctor into the main graph
This fused the KVCacheCollection constructor into the main graph,
removing the need to call a subgraph in order to construct the
collection object. This is expected to provide a minor perf improvement
in the generation loop.

NN_ORIG_REV_ID: 4ecdf7c8c61518c1dcf4f664dc85b951c083c050