[KGEN] Add hidden `mojo build` flag for Python gen (#49155)
Add a *hidden* flag to the `build` subcommand of `mojo` to control
whether to generate Python bindings. If the flag is specified, instead
of using the linker to produce an executable, it will produce a dynamic
library that contains the generated Python bindings for the user's
source input mojo file. This works out nicely since the Python bindings
generation in the parser all goes through `LIT::importMojoFile` where we
can just plumb this boolean flag through.

The output name of the produced dynamic library is still controlled via
the `-o` argument to `mojo build`. If not specified, it will now default
to the stem with the required suffix for importing Python modules
(`.so`) instead of no suffix as in the executable case on non-Windows
platforms.

In the future, we may do several things in this area such as:
- Make the delineation between producing an executable vs. dynamic
library more general from the UX side of `mojo build`.
- Allow generating a dynamic library output without Python bindings.
Right now, they are coupled, but there's no specific reason to do it
this way other than convenience for our current product-facing use case.
The way we spell this will change in the coming months as we get ready
to launch Mojo <-> Python interop.

Fixes MSTDL-975.

MODULAR_ORIG_COMMIT_REV_ID: b648f1588621b20df47611a9b1364068ac68d4aa