[KGEN][MOGGPreElab] Add @uses_opaque Decorator (#49292)
Opaque types in the graph compiler at this require access to the underlying constructor/destructor methods. Compared to tensors, regular mojo structs are instead reference counted by the runtime. When the reference reaches zero, the pointed to destructor is called.

Right now, this information for some reason is stored when parsing Mojo code. Specifically, if the function is under a special decorator @mogg_register. We extend this behavior for now to also cover a new decorator @uses_opaque. When a function is written with this, additional information is added to the LIT/KGEN signature.

This is needed to lower opaque types in a future PR under the new API; it would be weird to see @mogg_register in the new API!

REGISTER_ORIG_REV_ID: 79520d30af97c90b03385a849b14315cc8b0a153