[NFC][KGEN][MOGG] Remove `initialize_output` (#49296)
This was an old design to separate all the oddity with KGEN calling conventions to a single method. However we've bitten the bullet and supported opaque types in `execute`. We will try to manage the complexity in other means internally within the codebase.

NFC since this was not really used anywhere.

MOGG_ORIG_REV_ID: 2bd20495ec33956a40171a03a5415a89d2900566