[FileSystem] Default to cwd for non-existent files (#47913)
When determining the available space to open a file, we get the parent
of the file. However, the parent may not exist for files that the user
expects to be created. In those cases, we can default to the current
working directory since the files without a parent will be created
there.
Closes MSDK-1034

Signed-off-by: Rashid Kaleem <rashid@modular.com>
MODULAR_ORIG_COMMIT_REV_ID: 066b8423afd77f68446315defa43c83f4a282b33