[vscode] Don't restart the extension when the active sdk changes (#49629)
We've been following an approach in which the extension would restart
when the active sdk changes. This won't play well with our multi-sdk
project, as we only want to restart the functionalities corresponding to
only the projects whose SDK changed.
Instead, in this PR I'm removing the restart-after-sdk logic and adding
a listener in the LSP so that only the LSP restarts when the active SDK
changes. In fact, the LSP s the only feature that is long lived.

Towards MOTO-789

MODULAR_ORIG_COMMIT_REV_ID: fa4c91527eeb3072f2e9bfcfed5dbb4e1a089166