[KGEN] Plumb emission options through the compile_assembly (#49906)
This allows one to specify emission options to the compile_assembly
param operator. This allow one to control what gets passed into the
llvm:cl options. We want this because we want to be able to pass
nvptx-short-ptr=true to the compiler.

The hope was [Internal link] would have
been sufficient, but as noted in
[Internal link]
it is not and one has to plumb the cl option through.

Towards MOCO-1390

---------

Co-authored-by: Chris Lattner <clattner@nondot.org>
COMPILE_ORIG_REV_ID: 7c4049583e69894ac3affd40939795a576000000