[KGEN] Fix bug in division attribute simplification (#50155)
The KGEN attribute simplification for the division operation makes use
of llvm::GreatestCommonDenominator has some surprising behavior when
used with negative numbers. For instance

    GreatestCommonDenominator(-4, -1) == 3

This is likely due to GreatestCommonDenominator not knowing the
signedness of the value it is representing. The fix is to take the
absolute value when operating on signed integers, and separately factor
out the additional -1 terms if both inputs are negative.

MODULAR_ORIG_COMMIT_REV_ID: ba0e41afcbde08431b8e8c3704b5548bfb8b1745