[Kernels][GPU] Implements lane_id for AMD GPUs (#50173)
Define the `lane_id` for AMD GPUs as
```
threadIdx.x() & (WARP_SIZE - 1)
````
We should consider this for NVidia as well as it might be faster than
reading from a special register as is done now. but this should be
benchmarked before we make the change.

[Internal link]

GPU_ORIG_REV_ID: 806a48a070c8cda7b5f7b4bec90a0ba8859696da