[KGEN] strip-file-prefix flag (#49966)
Fixes [MOCO-1378: Allow storing relative paths for debug
[Internal link]

New compilation flag `-strip-file-prefix=<x>`. When non-empty, any files
with paths that have `<x>` as a prefix will have that be removed. Files
not under `<x>` will be left alone. This is similar to the behavior in
clang. This flag is added to `kgen`, `kgen-translate`, and `mojo`.

This applies to all MLIR Locations created in the parser & in Diags. In
particular, this flag is owned by `M::Diags`. There are typically two
ways locations are created during parsing:
- Vast majority of cases: Translating a SMLoc into a Location.
- Minority of cases: Constructed directly by providing a filename
directly.

Diags already handles the first case, so it can perform this
prefix-checking when building a location (this PR just adds a lookup
cache to speed up buffer-id to filename). For the second case, we expose
a function from Diags to "canonicalize" the filename, and the callers
are expected to use this as the filename when creating Locations or
DIFiles.

Caveat 1: The use of `MemoryBuffer::getBufferIdentifier` as filenames is
not totally aligned with its intended semantics, even though that's how
it's done basically everywhere. It breaks down when the MemoryBuffer is
backed not by a file, but a separate buffer somewhere. This is used by
-split-input-file & repl expressions. If we didn't care about the path
`getBufferIdentifier` returned, and just blindly stuffed it into a mlir
Location, that's fine. But if we want to treat it like a path and expand
it into an absolute path, we immediately run into the non-file-backed
buffer problem, where the names aren't paths, and expanding it will
create nonsense. To handle this, Diags does not attempt to expand the
path we get from `getBufferIdentifier` at all. It treats it as an opaque
string, and just blindly tries to remove the prefix. This means the
burden is on the user creating the MemoryBuffers to always open files
using their absolute paths if they want to match a `fileCompilationDir`
that is also an absolute path. Since the user is the one opening files
and also passing in `fileCompilationDir`, at least all this logic is
delegated to one owner, so they choose what format they want and what to
strip.

Caveat 2: In general, we should not dereference symlinks when traversing
directories for import, since the user may have some sort of setup where
they explicitly mapped the file elsewhere, bazel being one prominent
example. `std::filesystem::canonical` will resolve symlinks, and I had
to change to `absolute` to avoid this behavior so that tests also work
in bazel (they always work outside of bazel).

MODULAR_ORIG_COMMIT_REV_ID: 0498c4dc06e0eac0052333c0b49cc48fc6b195bf